Changed constructor to take double instead of float

+++ Core/ConveyorItem.cs
@@ -8,7 +8,7 @@ public class ConveyorItem
     public ItemID ItemID = 0;
     // distance along the conveyor in tiles
     public double Progress = 0;
-    public ConveyorItem(ItemID id, float progress)
+    public ConveyorItem(ItemID id, double progress)
     {
         ItemID = id;
         Progress = progress;
